LPS-192817 Moves search container classes
