+++ modules/apps/organizations/organizations-api/src/main/java/com/liferay/organizations/search/OrganizationSearchTerms.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portlet.usersadmin.search;
+package com.liferay.organizations.search;
 
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.dao.search.DAOParamUtil;
