+++ modules/apps/roles/roles-admin-api/src/main/java/com/liferay/roles/admin/search/GroupRoleChecker.java
@@ -3,13 +3,12 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portlet.rolesadmin.search;
+package com.liferay.roles.admin.search;
 
 import com.liferay.portal.kernel.dao.search.EmptyOnClickRowChecker;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.model.Organization;
 import com.liferay.portal.kernel.model.Role;
 import com.liferay.portal.kernel.service.GroupLocalServiceUtil;
 
@@ -18,9 +17,9 @@ import javax.portlet.RenderResponse;
 /**
  * @author Brian Wing Shun Chan
  */
-public class OrganizationRoleChecker extends EmptyOnClickRowChecker {
+public class GroupRoleChecker extends EmptyOnClickRowChecker {
 
-	public OrganizationRoleChecker(RenderResponse renderResponse, Role role) {
+	public GroupRoleChecker(RenderResponse renderResponse, Role role) {
 		super(renderResponse);
 
 		_role = role;
@@ -28,11 +27,9 @@ public class OrganizationRoleChecker extends EmptyOnClickRowChecker {
 
 	@Override
 	public boolean isChecked(Object object) {
-		Organization organization = (Organization)object;
+		Group group = (Group)object;
 
 		try {
-			Group group = organization.getGroup();
-
 			return GroupLocalServiceUtil.hasRoleGroup(
 				_role.getRoleId(), group.getGroupId());
 		}
@@ -45,13 +42,13 @@ public class OrganizationRoleChecker extends EmptyOnClickRowChecker {
 
 	@Override
 	public boolean isDisabled(Object object) {
-		Organization organization = (Organization)object;
+		Group group = (Group)object;
 
-		return isChecked(organization);
+		return isChecked(group);
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
-		OrganizationRoleChecker.class);
+		GroupRoleChecker.class);
 
 	private final Role _role;
