+++ modules/apps/roles/roles-admin-api/src/main/java/com/liferay/roles/admin/search/UserGroupRoleChecker.java
@@ -3,23 +3,24 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portlet.rolesadmin.search;
+package com.liferay.roles.admin.search;
 
 import com.liferay.portal.kernel.dao.search.EmptyOnClickRowChecker;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.Role;
+import com.liferay.portal.kernel.model.UserGroup;
 import com.liferay.portal.kernel.service.GroupLocalServiceUtil;
 
 import javax.portlet.RenderResponse;
 
 /**
- * @author Brian Wing Shun Chan
+ * @author Charles May
  */
-public class GroupRoleChecker extends EmptyOnClickRowChecker {
+public class UserGroupRoleChecker extends EmptyOnClickRowChecker {
 
-	public GroupRoleChecker(RenderResponse renderResponse, Role role) {
+	public UserGroupRoleChecker(RenderResponse renderResponse, Role role) {
 		super(renderResponse);
 
 		_role = role;
@@ -27,11 +28,13 @@ public class GroupRoleChecker extends EmptyOnClickRowChecker {
 
 	@Override
 	public boolean isChecked(Object object) {
-		Group group = (Group)object;
+		UserGroup userGroup = (UserGroup)object;
 
 		try {
+			Group userGroupGroup = userGroup.getGroup();
+
 			return GroupLocalServiceUtil.hasRoleGroup(
-				_role.getRoleId(), group.getGroupId());
+				_role.getRoleId(), userGroupGroup.getGroupId());
 		}
 		catch (Exception exception) {
 			_log.error(exception);
@@ -42,13 +45,13 @@ public class GroupRoleChecker extends EmptyOnClickRowChecker {
 
 	@Override
 	public boolean isDisabled(Object object) {
-		Group group = (Group)object;
+		UserGroup userGroup = (UserGroup)object;
 
-		return isChecked(group);
+		return isChecked(userGroup);
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
-		GroupRoleChecker.class);
+		UserGroupRoleChecker.class);
 
 	private final Role _role;
