+++ modules/apps/users-admin/users-admin-api/src/main/java/com/liferay/users/admin/search/UserSearch.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portlet.usersadmin.search;
+package com.liferay.users.admin.search;
 
 import com.liferay.portal.kernel.dao.search.SearchContainer;
 import com.liferay.portal.kernel.log.Log;
